Add multiplatform support and fix ARM x32 building
- Fix prefix_to_netmask_v4 overflow when prefix is 0 (shift left overflow)
- Add comprehensive tests for netmask conversion functions
- Improve ioctl wrapper for cross-platform compatibility:
  - Handle musl vs glibc differences
  - Proper 32-bit vs 64-bit pointer width handling
  - Support ARM32, ARMv7, ARM64, x86, x86_64, RISC-V64
- Add platform detection constant for debugging
- Fix clippy warnings:
  - Rename from_str to parse to avoid trait confusion
  - Fix needless_range_loop warnings
  - Fix explicit_auto_deref warning
  - Fix redundant_closure warning
- Update Cargo.toml with cross-compilation documentation
- Bump version to 0.3.3

+++ src/network.rs
@@ -295,13 +295,13 @@ impl Default for EventLoop {
 }
 
 // Poll constants
-pub const POLLIN: i16 = libc::POLLIN as i16;
+pub const POLLIN: i16 = libc::POLLIN;
 #[allow(dead_code)]
-pub const POLLOUT: i16 = libc::POLLOUT as i16;
+pub const POLLOUT: i16 = libc::POLLOUT;
 #[allow(dead_code)]
-pub const POLLERR: i16 = libc::POLLERR as i16;
+pub const POLLERR: i16 = libc::POLLERR;
 #[allow(dead_code)]
-pub const POLLHUP: i16 = libc::POLLHUP as i16;
+pub const POLLHUP: i16 = libc::POLLHUP;
 
 // ═══════════════════════════════════════════════════════════════════════════
 // HELPER FUNCTIONS
